pyproject.toml
[project]
name = "openalex-etl"
version = "0.1.0"
description = "This project is build to extarct and clean and load the openalex finance papers metadata to database for usage and analysis"
authors = [
    {name = "Gowtham",email = "gowthamgoli02@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "pandas (>=2.3.0,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "fsspec (==2024.3.1)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "boto3 (>=1.38.37,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pyodbc (>=5.2.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

